# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: PredictIQ API
  version: 1.0.0
  description: |
    REST API for the PredictIQ prediction markets platform.

    ## API Versioning

    The API uses URL path versioning (`/api/v1/`). The current stable version is **v1**.

    Clients may also send an `API-Version` header (e.g. `API-Version: v1`) to explicitly
    declare the version they target. If omitted, the server defaults to the current version.

    ## Deprecation Policy

    When a version is deprecated:
    - Responses will include a `Deprecation` header set to `true`.
    - A `Sunset` header will indicate the date after which the version will be removed.
    - A `Link` header will point to migration documentation.

    Clients should monitor these headers and migrate before the sunset date.

    Deprecated versions are supported for a minimum of **12 months** after the deprecation
    announcement before being removed.
servers:
  - url: http://0.0.0.0:8080
    description: Local development (default bind 0.0.0.0:8080)
...
paths:
...
  /webhooks/sendgrid:
    post:
      tags: [webhooks]
      operationId: sendgridWebhook
      summary: SendGrid event webhook receiver
      security:
        - SendGridWebhook: []
      parameters:
        - name: X-Twilio-Email-Event-Webhook-Signature
          in: header
          required: true
          schema:
            type: string
          description: HMAC-SHA256 signature of timestamp+body, base64-encoded
        - name: X-Twilio-Email-Event-Webhook-Timestamp
          in: header
          required: true
          schema:
            type: string
          description: Unix timestamp of the request (replay protection window is 300 s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AnyObject"
      responses:
        "200":
          description: Webhook processed
        "401":
          $ref: "#/components/responses/ApiError"
components:
...
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error description
...
  responses:
    ApiError:
      description: Error response with machine-readable code
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key required for protected endpoints
    SendGridWebhook:
      type: apiKey
      in: header
      name: X-Twilio-Email-Event-Webhook-Signature
      description: HMAC-SHA256 signature provided by SendGrid on webhook delivery